Survex 1.4.21

# -----BEGIN PGP SIGNATURE-----
#
# iQIzBAABCgAdFiEECOJAD/f+j+3jrLUoGBR7BzutKwcFAmoNJB4ACgkQGBR7Bzut
# KwdTVxAA0GtlNfNqhiOsWwBesne5FmCjotoU9KkjfH0WtE2I7PnxKvwTqEKuQw9V
# 1KaX8OKYN+1JndmRAjLLoLwNe5mHiGQWQ/3iluSEuM3H5IJddA4VXhfEoUuvrY2a
# 7mNfAh54Km07OjF6fx4Ba93CDJg3yspb+egiJiE4mg/mjpfqBYoDEHMfN01nRQ2d
# fcwcwnlMqSXOgXbKogdBPCCTe3I5miUIbtEMUtdBdTrWAAW7CsYHPmkD/8n8TKTf
# LsED4bws7oyqa7fBxAtg/ZyhG6vgo3yEZmV/tKs8/+sZGBW0X+4Boqd5bTlhYz3A
# PYj75a35OWwBYi3444CnoviI6hIUZ84o0bSSr9DsrbUEVgp6Jh9dZ87SbFNUGYcM
# 44tIkLeOz7BOTVoH1byfJN0xPhwm+3qkKJkfY9UpXhWpiR1/5qTzNaZVvpjMF5GZ
# L3NEqMoDJYlXIAXEwYGz+9eg5svZi7kDKIRVbn4VUP3xCtkge5hNU2LbXknKYcFG
# fLyE4HARZrzVry2fHFbzVtTDN6AtKoRRnFPP22laNANp2tDfbMgguwCwR3s6SNz2
# eDrtmQQ0m/ze5O1WmVSzw3h7En+VLNoVL20dPfNEGRkwMY1RN3H7WBJHtADe6neS
# AIAn9lt8rZ0e4BW46WnQlIb9bBSE3U3L1rwZEiRKvwwwwg3AFBc=
# =Z/2F
# -----END PGP SIGNATURE-----
# gpg: Signature made Tue May 19 22:01:50 2026 CDT
# gpg:                using RSA key 08E2400FF7FE8FEDE3ACB52818147B073BAD2B07
# gpg: Can't check signature: No public key

# Conflicts:
#	src/aboutdlg.h
#	src/avenprcore.h
#	src/aventreectrl.h
#	src/cavernlog.h
#	src/export.h
#	src/exportfilter.h
#	src/getopt.c
#	src/gla.h
#	src/glbitmapfont.h
#	src/guicontrol.h
#	src/log.h
#	src/mainfrm.h
#	src/model.h

+++ tests/diffpos.tst
@@ -1,7 +1,7 @@
 #!/bin/sh
 #
 # Survex test suite - diffpos tests
-# Copyright (C) 1999-2024 Olly Betts
+# Copyright (C) 1999-2025 Olly Betts
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -14,8 +14,8 @@
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+# along with this program; if not, see
+# <https://www.gnu.org/licenses/>.
 
 testdir=`echo $0 | sed 's!/[^/]*$!!' || echo '.'`
 
@@ -76,11 +76,9 @@ for file in $TESTS ; do
     fi
     rm "$vg_log"
   fi
-  if test -n "$VERBOSE" ; then
-    cat diffpos.tmp
-    $DIFF diffpos.tmp "$srcdir/${file}.out" || exit 1
-  else
-    $QUIET_DIFF diffpos.tmp "$srcdir/${file}.out" > /dev/null || exit 1
+  if ! $QUIET_DIFF diffpos.tmp "$srcdir/${file}.out" > /dev/null ; then
+    [ -z "$VERBOSE"] || $DIFF diffpos.tmp "$srcdir/${file}.out"
+    exit 1
   fi
   rm -f diffpos.tmp
 done
@@ -89,10 +87,14 @@ for args in '' '--survey survey' '--survey survey.xyzzy' '--survey xyzzy' ; do
   echo "diffpos $args"
   rm -f diffpos.tmp
   $DIFFPOS $args "$srcdir/v0.3d" "$srcdir/v0.3d" > diffpos.tmp
-  if test -n "$VERBOSE" ; then
-    cat diffpos.tmp
+  exitcode=$?
+  if [ $exitcode != 0 ] ; then
+    [ -z "$VERBOSE"] || cat diffpos.tmp
   fi
-  $QUIET_DIFF diffpos.tmp /dev/null > /dev/null || exit 1
+  if ! $QUIET_DIFF diffpos.tmp /dev/null > /dev/null ; then
+    [ -z "$VERBOSE"] || $DIFF diffpos.tmp /dev/null
+    exit 1
+  fi
   rm -f diffpos.tmp
 done
 test -n "$VERBOSE" && echo "Test passed"
